+++ src/factories.py
@@ -5,20 +5,15 @@
 
 import logging
 
-from github_runner_manager import constants
 from github_runner_manager.configuration import (
     ApplicationConfiguration,
     Flavor,
     Image,
     NonReactiveCombination,
     NonReactiveConfiguration,
-    QueueConfig,
-    ReactiveConfiguration,
     SupportServiceConfig,
-    UserInfo,
 )
 from github_runner_manager.configuration.github import GitHubConfiguration
-from github_runner_manager.manager.runner_scaler import RunnerScaler
 from github_runner_manager.openstack_cloud.configuration import (
     OpenStackConfiguration,
     OpenStackCredentials,
@@ -27,26 +22,6 @@ from github_runner_manager.openstack_cloud.configuration import (
 from charm_state import CharmState
 
 logger = logging.getLogger(__name__)
-
-
-def create_runner_scaler(state: CharmState, app_name: str, unit_name: str) -> RunnerScaler:
-    """Get runner scaler instance for scaling runners.
-
-    Args:
-        state: The CharmState.
-        app_name: Name of the application.
-        unit_name: Unit name for the prefix for instances.
-
-    Returns:
-        An instance of RunnerScaler.
-    """
-    application_configuration = create_application_configuration(state, app_name, unit_name)
-    user = UserInfo(constants.RUNNER_MANAGER_USER, constants.RUNNER_MANAGER_GROUP)
-
-    return RunnerScaler.build(
-        application_configuration=application_configuration,
-        user=user,
-    )
 
 
 def create_application_configuration(
@@ -84,7 +59,6 @@ def create_application_configuration(
         custom_pre_job_script=state.charm_config.custom_pre_job_script,
     )
     non_reactive_configuration = _get_non_reactive_configuration(state)
-    reactive_configuration = _get_reactive_configuration(state, app_name)
     openstack_configuration = create_openstack_configuration(state, unit_name)
     return ApplicationConfiguration(
         allow_external_contributor=state.charm_config.allow_external_contributor,
@@ -93,7 +67,6 @@ def create_application_configuration(
         github_config=github_configuration,
         service_config=service_config,
         non_reactive_configuration=non_reactive_configuration,
-        reactive_configuration=reactive_configuration,
         openstack_configuration=openstack_configuration,
         planner_url=state.planner_config.endpoint if state.planner_config else None,
         planner_token=state.planner_config.token if state.planner_config else None,
@@ -133,43 +106,6 @@ def _get_non_reactive_configuration(state: CharmState) -> NonReactiveConfigurati
             )
         ]
     return NonReactiveConfiguration(combinations=combinations)
-
-
-def _get_reactive_configuration(state: CharmState, app_name: str) -> ReactiveConfiguration:
-    """Get ReactiveConfiguration from CharmState and app_name.
-
-    Currently only one image and one flavor is supported.
-    """
-    if not state.reactive_config:
-        return None
-    reactive_config = state.reactive_config
-    openstack_image = state.runner_config.openstack_image
-    image_labels = []
-    images = []
-    flavors = []
-    if openstack_image and openstack_image.id:
-        if openstack_image.tags:
-            image_labels = openstack_image.tags
-        image = Image(
-            name=openstack_image.id,
-            labels=image_labels,
-        )
-        flavor = Flavor(
-            name=state.runner_config.flavor_label_combinations[0].flavor,
-            labels=(
-                []
-                if not state.runner_config.flavor_label_combinations[0].label
-                else [state.runner_config.flavor_label_combinations[0].label]
-            ),
-        )
-        images = [image]
-        flavors = [flavor]
-    return ReactiveConfiguration(
-        queue=QueueConfig(mongodb_uri=reactive_config.mq_uri, queue_name=app_name),
-        max_total_virtual_machines=state.runner_config.max_total_virtual_machines,
-        images=images,
-        flavors=flavors,
-    )
 
 
 def create_openstack_configuration(state: CharmState, unit_name: str) -> OpenStackConfiguration: